Backend/requirements.txt
# Core Framework
fastapi==0.115.12
uvicorn==0.24.0
python-dotenv==1.1.0
pydantic
aiohttp==3.9.3
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
prometheus-client==0.21.1
structlog==24.1.0
httpx==0.28.1
mcp==1.6.0
starlette==0.46.2
strawberry-graphql[fastapi]

# AI and Machine Learning
sentence-transformers==2.2.2
protobuf==3.20.3
numpy==1.26.4
scipy==1.11.4
pandas==2.1.4
scikit-learn==1.4.0
huggingface-hub==0.19.4
openai==1.64.0
langchain
langchain-core
langchain-openai
langchain-community

# Database
sqlalchemy==1.4.41
alembic==1.14.1
asyncpg==0.30.0
aiosqlite==0.21.0

# MongoDB
pymongo==4.6.1
motor==3.3.2

# Utilities
cryptography==44.0.0

PyJWT==2.8.0
redis==5.0.4

chromadb==0.4.22

atomic-agents==1.1.3

aioredis
gql[all]

dnspython